Board
Lots of progress with the PCB, size increased to 1.25 X 3.5 inches to
better fit circuitry and allow for an adequate keep out layer. Expansion
connectors are being added but highly subject to change a few times
before finalizing. The ESP-02 module's UART signals are jumperable to
connect it to either the STM32 or a programming pin header. Discussing
with Shane about further improvements and ESD suppression.